pyproject.toml
[project]
name = "de-zoomcamp-hw7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
]
